+++ testsuite/test_tupletree.py
@@ -5,12 +5,13 @@
 """
 import os
 import xml
+import pytest
 import unittest
 import pprint
 import six
 from pkg_resources import resource_filename
 
-from pywbem import tupletree
+from pywbem import tupletree, ParseError
 
 pp = pprint.PrettyPrinter(indent=4)  # pylint: disable=invalid-name
 
@@ -197,7 +198,7 @@ class TestTupleTreeRegression(unittest.TestCase):
             with open(fn, 'rb') as fh:
                 xml_str = fh.read()
             tree_dom = xml_to_tupletree(xml_str)
-            tree_sax = tupletree.xml_to_tupletree_sax(xml_str)
+            tree_sax = tupletree.xml_to_tupletree_sax(xml_str, 'Test XML')
             self.assertEqual(tree_dom, tree_sax,
                              'SAX and DOM not equal for %s:\n%s,\n%s'
                              % (fn, pp.pformat(tree_sax), pp.pformat(tree_dom)))
@@ -216,8 +217,9 @@ class TestTupleTreeError(unittest.TestCase):
         filename = os.path.join(self.data_dir, 'Associators_error.xml')
         with open(filename, 'rb') as fh:
             xml_str = fh.read()
-            self.assertRaises(xml.sax.SAXParseException,
-                              tupletree.xml_to_tupletree_sax, xml_str)
+            self.assertRaises(ParseError,
+                              tupletree.xml_to_tupletree_sax,
+                              xml_str, 'Test XML')
 
 
 class TestTupleTreeSax(unittest.TestCase):
@@ -232,10 +234,233 @@ class TestTupleTreeSax(unittest.TestCase):
         path = os.path.join(data_dir, 'Associators_StorageVolume_small.xml')
         with open(path, 'rb') as fh:
             xml_str = fh.read()
-        tree_sax = tupletree.xml_to_tupletree_sax(xml_str)
+        tree_sax = tupletree.xml_to_tupletree_sax(xml_str, 'Test XML')
         self.assertEqual(test_tuple, tree_sax,
                          'Test tuple and SAX not equal for %s:\n%s,\n%s'
                          % (path, pp.pformat(tree_sax), pp.pformat(tree_sax)))
+
+
+class Test_check_invalid_utf8_sequences(object):
+    """Tests for check_invalid_utf8_sequences()"""
+
+    testcases = [
+        # Each testcase has these items:
+        # * desc: Short testcase description.
+        # * utf8_string: Input string for the function.
+        # * exp_exc_type: Expected exception type, or None if expecting success.
+        # * condition: Condition for testcase to run.
+
+        # General cases
+        (
+            "Incorrect type of input string",
+            u'<V>ab</V>',
+            TypeError, True
+        ),
+
+        # Good cases
+        (
+            "Good case without special chars",
+            b'<V>a</V>',
+            None, True
+        ),
+        (
+            "Good case with U+0009 (TAB), U+000A (NL), U+000D (CR) chars",
+            b'<V>a\x09b\x0Ac\x0Dd</V>',
+            None, True
+        ),
+        (
+            "Good case with U+0350 char (in UTF-8)",
+            b'<V>a\xCD\x90b</V>',
+            None, True
+        ),
+        (
+            "Good case with U+2013 char (in UTF-8)",
+            b'<V>a\xE2\x80\x93b</V>',
+            None, True
+        ),
+        (
+            "Good case with U+10122 char (in UTF-8)",
+            b'<V>a\xF0\x90\x84\xA2b</V>',
+            None, True
+        ),
+
+        # Correctly encoded but ill-formed UTF-8
+        (
+            "Ill-formed UTF-8 using surrogate U+D800,U+DD22 "
+            "(in otherwise correctly encoded UTF-8)",
+            b'<V>a\xED\xA0\x80\xED\xB4\xA2b</V>',
+            ParseError, True
+        ),
+        (
+            "Ill-formed UTF-8 using surrogates U+D800,U+DD22 and U+D800,U+DD23 "
+            "(in otherwise correctly encoded UTF-8)",
+            b'<V>a\xED\xA0\x80\xED\xB4\xA2b\xED\xA0\x80\xED\xB4\xA3</V>',
+            ParseError, True
+        ),
+
+        # Incorrectly encoded UTF-8
+        (
+            "Incorrectly encoded UTF-8 with 1-byte sequence",
+            b'<V>a\x80b</V>',
+            ParseError, True
+        ),
+        (
+            "Incorrectly encoded UTF-8 with 2-byte sequence "
+            "with missing second byte",
+            b'<V>a\xC0',
+            ParseError, True
+        ),
+        (
+            "Incorrectly encoded UTF-8 with 2-byte sequence "
+            "with incorrect 2nd byte",
+            b'<V>a\xC0b</V>',
+            ParseError, True
+        ),
+        (
+            "Incorrectly encoded UTF-8 with 4-byte sequence "
+            "with incorrect 3rd byte",
+            b'<V>a\xF1\x80abc</V>',
+            ParseError, True
+        ),
+        (
+            "Incorrectly encoded UTF-8 with 4-byte sequence "
+            "with incorrect 3rd byte that is an incorrect new start",
+            b'<V>a\xF1\x80\xFFbc</V>',
+            ParseError, True
+        ),
+        (
+            "Incorrectly encoded UTF-8 with 4-byte sequence "
+            "with incorrect 3rd byte that is a correct new start",
+            b'<V>a\xF1\x80\xC2\x81c</V>',
+            ParseError, True
+        ),
+    ]
+
+    @pytest.mark.parametrize(
+        "desc, utf8_string, exp_exc_type, condition",
+        testcases
+    )
+    def test_check_invalid_utf8_sequences(
+            self, desc, utf8_string, exp_exc_type, condition):
+
+        if not condition:
+            pytest.skip("Condition for test case not met")
+
+        if exp_exc_type is None:
+
+            unicode_string = tupletree.check_invalid_utf8_sequences(
+                utf8_string, "Test XML")
+
+            assert isinstance(unicode_string, six.text_type)
+            utf8_string_u = utf8_string.decode('utf-8')
+            assert unicode_string == utf8_string_u
+
+        else:
+            with pytest.raises(exp_exc_type):
+
+                tupletree.check_invalid_utf8_sequences(utf8_string, "Test XML")
+
+
+class Test_check_invalid_xml_chars(object):
+    """Tests for check_invalid_xml_chars()"""
+
+    testcases = [
+        # Each testcase has these items:
+        # * desc: Short testcase description.
+        # * xml_string: Input string for the function.
+        # * exp_exc_type: Expected exception type, or None if expecting success.
+        # * condition: Condition for testcase to run.
+
+        # General cases
+        (
+            "Incorrect type of input string",
+            b'<V>ab</V>',
+            TypeError, True
+        ),
+
+        # Good cases
+        (
+            "Good case without special chars",
+            u'<V>a</V>',
+            None, True
+        ),
+        (
+            "Good case with U+0009 (TAB), U+000A (NL), U+000D (CR) chars",
+            u'<V>a\u0009b\u000Ac\u000Dd</V>',
+            None, True
+        ),
+        (
+            "Good case with U+0350 char",
+            u'<V>a\u0350b</V>',
+            None, True
+        ),
+        (
+            "Good case with U+2013 char",
+            u'<V>a\u2013b</V>',
+            None, True
+        ),
+        (
+            "Good case with U+10122 char",
+            u'<V>a\u010122b</V>',
+            None, True
+        ),
+
+        # Invalid XML characters
+        (
+            "Invalid XML char U+0008 (BEL)",
+            u'<V>a\u0008b</V>',
+            ParseError, True
+        ),
+        (
+            "Invalid XML char U+0000",
+            u'<V>a\u0000b</V>',
+            ParseError, True
+        ),
+        (
+            "Invalid XML chars U+0001 ... U+0008",
+            u'<V>a\u0001b\u0002c\u0003d\u0004e\u0005f\u0006g\u0007h\u0008i</V>',
+            ParseError, True
+        ),
+        (
+            "Invalid XML chars U+000B ... U+000C",
+            u'<V>a\u000Bb\u000Cc</V>',
+            ParseError, True
+        ),
+        (
+            "Invalid XML chars U+000E ... U+0015",
+            u'<V>a\u000Eb\u000Fc\u0010d\u0011e\u0012f\u0013g\u0014h\u0015i</V>',
+            ParseError, True
+        ),
+        (
+            "Invalid XML chars U+0016 ... U+001D",
+            u'<V>a\u0016b\u0017c\u0018d\u0019e\u001Af\u001Bg\u001Ch\u001Di</V>',
+            ParseError, True
+        ),
+        (
+            "Invalid XML chars U+001E ... U+001F",
+            u'<V>a\u001Eb\u001Fc</V>',
+            ParseError, True
+        ),
+    ]
+
+    @pytest.mark.parametrize(
+        "desc, xml_string, exp_exc_type, condition",
+        testcases
+    )
+    def test_check_invalid_xml_chars(
+            self, desc, xml_string, exp_exc_type, condition):
+
+        if not condition:
+            pytest.skip("Condition for test case not met")
+
+        if exp_exc_type is None:
+
+            tupletree.check_invalid_xml_chars(xml_string, "Test XML")
+
+        else:
+            with pytest.raises(exp_exc_type):
+
+                tupletree.check_invalid_xml_chars(xml_string, "Test XML")
 
 
 if __name__ == '__main__':